Activate case button for activate case
