+++ src/config.tsx
@@ -4,22 +4,3 @@ export const ROUTES = {
   NEWCASE: "/new-case",
   CASEINFO: "/case/:caseid",
 };
-
-export const CASE_STATUS_ENUM = {
-  NONE: "None",
-  ACTIVATED: "Activated",
-  PENDING: "Pending",
-  REJECTED: "Rejected",
-} as const;
-
-export const FILTERED_CASE_STATUS_ENUM = {
-  NONE: "none",
-  ACTIVATED: "activated",
-  PENDING: "pending",
-  REJECTED: "rejected",
-} as const;
-
-export const NAV_ITEMS_ENUM = {
-  DASHBOARD: "dashboard",
-  SECOND: "second",
-} as const;
