+++ src/pages/dashboard/casestatistics/casestatisticcard/mostfrequentscamcard.tsx
@@ -1,5 +1,5 @@
 import { Card } from "react-bootstrap";
-import { CaseDataType1 } from "../../types/types";
+import { CaseDataType1 } from "../../../../types/types";
 
 interface MostPopularScamCardProps {
   caseData: CaseDataType1[];
