+++ src/pages/dashboard/casestatistics/casestatisticcard/statusscamscard.tsx
@@ -1,6 +1,6 @@
 import { Card } from "react-bootstrap";
-import { CASE_STATUS_ENUM } from "../../config";
-import { CaseDataType1 } from "../../types/types";
+import { CASE_STATUS_ENUM } from "../../../../types/enums";
+import { CaseDataType1 } from "../../../../types/types";
 
 interface StatusScamCard {
   caseData: CaseDataType1[];
