+++ src/types/enums.tsx
@@ -1,0 +1,18 @@
+export const CASE_STATUS_ENUM = {
+  NONE: "None",
+  ACTIVATED: "Activated",
+  PENDING: "Pending",
+  REJECTED: "Rejected",
+} as const;
+
+export const FILTERED_CASE_STATUS_ENUM = {
+  NONE: "none",
+  ACTIVATED: "activated",
+  PENDING: "pending",
+  REJECTED: "rejected",
+} as const;
+
+export const NAV_ITEMS_ENUM = {
+  DASHBOARD: "dashboard",
+  SECOND: "second",
+} as const;
